Fix visibility of region on a full add for headless
